move utility fns to lib file
