+++ Cargo.toml
@@ -1,16 +1,15 @@
 [package]
-name = "hemingway"
-version = "0.5.1"
 authors = ["Jahziel Villasana <jahzielv@gmail.com>"]
-edition = "2018"
-readme = "README.md"
-homepage = "https://github.com/jahzielv/hemingway#hemingway"
-repository = "https://github.com/jahzielv/hemingway"
-keywords = ["cli", "RSS", "reader"]
 categories = ["command-line-utilities"]
-license = "MIT"
 description = "An economical RSS reader for your terminal."
-
+edition = "2018"
+homepage = "https://github.com/jahzielv/hemingway#hemingway"
+keywords = ["cli", "RSS", "reader"]
+license = "MIT"
+name = "hemingway"
+readme = "README.md"
+repository = "https://github.com/jahzielv/hemingway"
+version = "0.6.0"
 
 [lib]
 name = "hemlib"
@@ -21,14 +20,13 @@ name = "hem"
 path = "src/main.rs"
 
 [dependencies]
-structopt = "0.3.15"
-reqwest = "0.10.6"
-tokio = {version = "0.2.21", features = ["full"]}
-feed-rs = "0.3.0"
-serde_json = "1.0.55"
-serde = "1.0.112"
-itertools = "0.9.0"
-dirs = "2.0.2"
-chrono = "0.4.11"
 ansi_term = "0.12.1"
-error-chain = "0.12.2"
+chrono = "0.4.11"
+dirs = "2.0.2"
+feed-rs = "0.3.0"
+itertools = "0.9.0"
+reqwest = "0.10.6"
+serde = "1.0.112"
+serde_json = "1.0.55"
+structopt = "0.3.15"
+tokio = {version = "0.2.21", features = ["full"]}
