+++ src/main.rs
@@ -2,14 +2,12 @@ use ansi_term::Style;
 use chrono::offset::Utc;
 use chrono::DateTime;
 use feed_rs::parser;
-use hemlib::ProcessedFeed;
-use serde::{Deserialize, Serialize};
+use hemlib::{add_feed, find_config, list_feeds, rust_to_config, ConfigObj, ProcessedFeed};
+use reqwest::Client;
 use std::fs;
-use std::fs::File;
 use std::io::BufReader;
 use std::io::Read;
 use std::io::Write;
-use std::path::Path;
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
@@ -34,55 +32,10 @@ enum Cmd {
     },
     List,
 }
-#[derive(Debug, Serialize, Deserialize)]
-struct Feed {
-    uri: String,
-    last_accessed: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct ConfigObj {
-    feeds: Vec<Feed>,
-}
-
-fn find_config() -> std::path::PathBuf {
-    let homedir: std::path::PathBuf = dirs::home_dir().expect("no home dir");
-    //let path_to_config: &Path =
-    Path::new(&homedir).join(".hemrc")
-}
-
-fn config_to_rust() -> ConfigObj {
-    let config = fs::read_to_string(&find_config()).expect("reading config failed");
-    serde_json::from_str(&config).unwrap()
-}
-
-fn rust_to_config(content: &[u8]) {
-    let mut file = match File::create(find_config()) {
-        Err(why) => panic!("config file access failed: {}", why),
-        Ok(file) => file,
-    };
-    file.write_all(content)
-        .expect("Writing to .hemrc failed :(");
-}
-
-fn add_feed(feed: &str) {
-    let mut my_feeds: ConfigObj = config_to_rust();
-    my_feeds.feeds.push(Feed {
-        uri: feed.to_owned(),
-        last_accessed: Utc::now().to_rfc3339().to_owned(),
-    });
-    rust_to_config(serde_json::to_string(&my_feeds).unwrap().as_bytes());
-}
-
-fn list_feeds() {
-    let my_feeds: ConfigObj = config_to_rust();
-    for f in my_feeds.feeds {
-        println!("{}", f.uri);
-    }
-}
 
 async fn top<'a>(num: usize) -> Result<Vec<ProcessedFeed>, Box<dyn std::error::Error>> {
     let config_path = find_config();
+    let client = Client::new();
     let mut processed: Vec<ProcessedFeed> = Vec::new();
     let config = match fs::read_to_string(&config_path) {
         Ok(config) => config,
@@ -115,18 +68,23 @@ async fn top<'a>(num: usize) -> Result<Vec<ProcessedFeed>, Box<dyn std::error::E
     };
 
     for i in 0..config_obj.feeds.len() {
-        let resp = reqwest::get(&config_obj.feeds[i].uri).await?.text().await?;
+        let resp = client
+            .get(&config_obj.feeds[i].uri)
+            .send()
+            .await?
+            .text()
+            .await?;
         let feed = parser::parse(resp.as_bytes()).unwrap();
         let procfeed = {
             let title = feed.title.unwrap();
             let title_owned = title.content.to_owned();
 
             let entries = feed.entries.iter().enumerate();
-            let mut it = Vec::<String>::new();
+            let mut processed_items = Vec::<String>::new();
             for (j, e) in entries {
                 if j < num {
                     let e_title = e.title.as_ref().unwrap();
-                    it.push(format!(
+                    processed_items.push(format!(
                         "{} \n\t  {}\n",
                         Style::new().italic().paint(e_title.content.clone()),
                         e.id
@@ -136,7 +94,7 @@ async fn top<'a>(num: usize) -> Result<Vec<ProcessedFeed>, Box<dyn std::error::E
 
             ProcessedFeed {
                 title: title_owned,
-                items: it,
+                items: processed_items,
             }
         };
         processed.push(procfeed);
@@ -148,6 +106,7 @@ async fn top<'a>(num: usize) -> Result<Vec<ProcessedFeed>, Box<dyn std::error::E
 async fn process_feed<'a>() -> Result<Vec<ProcessedFeed>, Box<dyn std::error::Error>> {
     let config_path = find_config();
     let mut processed: Vec<ProcessedFeed> = Vec::new();
+    let client = Client::new();
     let config = match fs::read_to_string(&config_path) {
         Ok(config) => config,
         Err(e) => {
@@ -179,32 +138,22 @@ async fn process_feed<'a>() -> Result<Vec<ProcessedFeed>, Box<dyn std::error::Er
     };
 
     for i in 0..config_obj.feeds.len() {
-        let resp = reqwest::get(&config_obj.feeds[i].uri).await?.text().await?;
+        let resp = client
+            .get(&config_obj.feeds[i].uri)
+            .send()
+            .await?
+            .text()
+            .await?;
         let feed = parser::parse(resp.as_bytes()).unwrap();
         let last_accessed = DateTime::from(
             DateTime::parse_from_rfc3339(&config_obj.feeds[i].last_accessed).unwrap(),
         );
-        // let feed_updates = feed.updated;
-        // if feed_updates.is_none() {
-        //     eprintln!(
-        //         "{}: no update data available for this feed",
-        //         feed.title.unwrap().content
-        //     );
-        //     config_obj.feeds[i].last_accessed = Utc::now().to_rfc3339().to_owned();
-        //     continue;
-        // }
-        // let duration = last_accessed - feed_updates.unwrap();
-        // if duration.num_seconds() > 0 {
-        //     println!("{}: Nothing new here...", feed.title.unwrap().content);
-        //     config_obj.feeds[i].last_accessed = Utc::now().to_rfc3339().to_owned();
-        //     continue;
-        // }
         let procfeed = {
             let title = feed.title.unwrap();
             let title_owned = title.content.to_owned();
 
             let entries = feed.entries.iter().enumerate();
-            let mut it = Vec::<String>::new();
+            let mut processed_items = Vec::<String>::new();
             let mut entry_date;
             for (j, e) in entries {
                 if e.updated.is_none() {
@@ -213,19 +162,19 @@ async fn process_feed<'a>() -> Result<Vec<ProcessedFeed>, Box<dyn std::error::Er
                     entry_date = e.updated.unwrap();
                 }
                 let entry_duration = last_accessed - entry_date; //e.updated.unwrap();
-                if j < 10 && entry_duration.num_seconds() < 0 {
+                if j < 5 && entry_duration.num_seconds() < 0 {
                     let e_title = e.title.as_ref().unwrap();
-                    it.push(format!("{} \n\t  {}\n", e_title.content.clone(), e.id));
+                    processed_items.push(format!("{} \n\t  {}\n", e_title.content.clone(), e.id));
                 }
             }
 
-            if it.len() == 0 {
-                it.push(format!("Nothing new here..."));
+            if processed_items.len() == 0 {
+                processed_items.push(format!("Nothing new here..."));
             }
 
             ProcessedFeed {
                 title: title_owned,
-                items: it,
+                items: processed_items,
             }
         };
         processed.push(procfeed);
